PARTICIPATION ACTIVITY 7.21.1: Validate domain names with functions. Functions facilitate breaking down a large problem into a collection of smaller ones.
A top-level domain (TLD) name is the last part of an Internet domain name like .com in example.com. A core generic top-level domain (core gTLD) is a TLD that is either .com, .net, .org, or .info. A restricted top-level domain is a TLD that is either .biz, .name, or .pro. A second-level domain is a single name that precedes a TLD as in apple in apple.com

The following program repeatedly prompts for a domain name and indicates whether that domain name is valid and has a core gTLD. For this program, a valid domain name has a second-level domain followed by a TLD, and the second-level domain has these three characteristics:

Is 1-63 characters in length.
Contains only uppercase and lowercase letters or a dash.
Does not begin or end with a dash.
For this program, a valid domain name must contain only one period, such as apple.com, but not support.apple.com. The program ends when the user presses just the Enter key in response to a prompt.

Run the program. Note that a restricted gTLD is not recognized as such.
Change the program by writing an input function and adding the validation for a restricted gTLD. Run the program again.